LayoutTests/imported/w3c/web-platform-tests/css/css-grid/alignment/grid-alignment-implies-size-change-029-expected.txt

CONSOLE MESSAGE: line 190: No valid data-* attributes found in selector list : .after


PASS .before 1 
